MapScreen完成界面拆分原AMap更名为Navigation，新增Discovery和Routing两个Screen
